fix : delete extra code, change H2 BD to Postgres, add docker and docker compose files, add kit of base queries to filmorate-service for Postman, make structure of project clearer

+++ src/main/java/ru/yandex/practicum/filmorate/model/MPA.java
@@ -2,13 +2,12 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.Data;
 
-
 @Data
 public class MPA {
-    private int id;
+    private long id;
     private String name;
 
-    public MPA(int id, String name) {
+    public MPA(long id, String name) {
         this.id = id;
         this.name = name;
     }
